Fixes for issues introduced with renaming
