+++ py_cui/controls/__init__.py
@@ -1,1 +1,4 @@
-import py_cui.control_widgets.slider
+"""A collection of modules containing control-style widgets and popups, like sliders and gauges
+"""
+
+import py_cui.controls.slider
