+++ py_cui/dialogs/__init__.py
@@ -1,5 +1,5 @@
 """A collection of modules containing dialog-style widgets and popups.
 """
 
-import py_cui.dialog_widgets.form
-import py_cui.dialog_widgets.explorer
+import py_cui.dialogs.form
+import py_cui.dialogs.filedialog
